fix(cap): downcast to 184

+++ src/interfaces/IMetaMorpho.sol
@@ -79,10 +79,10 @@ interface IMetaMorphoBase {
 /// @dev This interface is inherited by MetaMorpho so that function signatures are checked by the compiler.
 /// @dev Consider using the IMetaMorpho interface instead of this one.
 interface IMetaMorphoStaticTyping is IMetaMorphoBase {
-    function config(Id) external view returns (uint192 cap, bool enabled, uint56 removableAt);
-    function pendingGuardian() external view returns (address guardian, uint56 validAt);
-    function pendingCap(Id) external view returns (uint192 value, uint56 validAt);
-    function pendingTimelock() external view returns (uint192 value, uint56 validAt);
+    function config(Id) external view returns (uint184 cap, bool enabled, uint64 removableAt);
+    function pendingGuardian() external view returns (address guardian, uint64 validAt);
+    function pendingCap(Id) external view returns (uint192 value, uint64 validAt);
+    function pendingTimelock() external view returns (uint192 value, uint64 validAt);
 }
 
 /// @title IMetaMorpho
